update: Show date fields with time in table row detaitls
